+++ lib/enums/date_range_type.dart
@@ -1,0 +1,27 @@
+/// Enum representing various date range types for easy date checks.
+enum DateRangeType {
+  today,
+  yesterday,
+  tomorrow,
+  thisWeek,
+  last7Days,
+  leapYear,
+
+  /// The current month
+  thisMonth,
+
+  /// The previous month
+  lastMonth,
+
+  /// The next month
+  nextMonth,
+
+  /// The current year
+  thisYear,
+
+  /// The previous year
+  lastYear,
+
+  /// The next year
+  nextYear,
+}